Update that allows to select a device
Hi!

I made a little update to this script that will print all the connected devices `name` and `id` to the console. You can then use a specific device by adding `device = <name|id>` to the code.

I also adjusted the normalization so it actually normalizes 0-1, but maybe you want to put this back to how it was. Previously it would not have 0 as the lowest value because a `1` was added before division.

Let me know what you think! :)